Merge base branch into test branch for ARM compatibility

+++ src/index.ts
@@ -1,21 +1,26 @@
-import { build, BuildResult, BuildOptions } from 'esbuild';
+import { build, BuildResult, BuildOptions, Plugin } from 'esbuild';
 import * as fs from 'fs-extra';
 import * as globby from 'globby';
 import * as path from 'path';
-import { mergeRight } from 'ramda';
+import * as pMap from 'p-map';
+import { concat, always, memoizeWith, mergeRight } from 'ramda';
 import * as Serverless from 'serverless';
-import * as Plugin from 'serverless/classes/Plugin';
+import * as ServerlessPlugin from 'serverless/classes/Plugin';
 import * as chokidar from 'chokidar';
 
-import { extractFileNames } from './helper';
+import { extractFileNames, providerRuntimeMatcher } from './helper';
 import { packExternalModules } from './pack-externals';
 import { pack } from './pack';
 import { preOffline } from './pre-offline';
 import { preLocal } from './pre-local';
+import { trimExtension } from './utils';
 
 export const SERVERLESS_FOLDER = '.serverless';
 export const BUILD_FOLDER = '.build';
 export const WORK_FOLDER = '.esbuild';
+
+type Plugins = Plugin[];
+type ReturnPluginsFn = (sls: Serverless) => Plugins;
 
 interface OptionsExtended extends Serverless.Options {
   verbose?: boolean;
@@ -26,37 +31,50 @@ export interface WatchConfiguration {
   ignore?: string[] | string;
 }
 
-export interface Configuration extends Omit<BuildOptions, 'watch'> {
+export interface PackagerOptions {
+  scripts?: string[] | string;
+}
+
+export interface Configuration extends Omit<BuildOptions, 'nativeZip' | 'watch' | 'plugins'> {
+  concurrency?: number;
   packager: 'npm' | 'yarn';
   packagePath: string;
-  exclude: string[];
+  exclude: '*' | string[];
+  nativeZip: boolean;
   watch: WatchConfiguration;
+  plugins?: string;
+  keepOutputDirectory?: boolean;
+  packagerOptions?: PackagerOptions;
 }
 
 const DEFAULT_BUILD_OPTIONS: Partial<Configuration> = {
   bundle: true,
-  target: 'es2017',
+  target: 'node10',
   external: [],
   exclude: ['aws-sdk'],
+  nativeZip: false,
   packager: 'npm',
   watch: {
     pattern: './**/*.(js|ts)',
     ignore: [WORK_FOLDER, 'dist', 'node_modules', SERVERLESS_FOLDER],
   },
+  keepOutputDirectory: false,
+  packagerOptions: {},
 };
 
-export class EsbuildPlugin implements Plugin {
+export class EsbuildServerlessPlugin implements ServerlessPlugin {
+  serviceDirPath: string;
   workDirPath: string;
   buildDirPath: string;
 
   serverless: Serverless;
   options: OptionsExtended;
-  hooks: Plugin.Hooks;
-  buildOptions: Configuration;
+  hooks: ServerlessPlugin.Hooks;
   buildResults: {
     result: BuildResult;
     bundlePath: string;
     func: any;
+    functionAlias: string;
   }[];
   packExternalModules: () => Promise<void>;
   pack: () => Promise<void>;
@@ -71,13 +89,11 @@ export class EsbuildPlugin implements Plugin {
     this.preOffline = preOffline.bind(this);
     this.preLocal = preLocal.bind(this);
 
-    this.workDirPath = path.join(this.serverless.config.servicePath, WORK_FOLDER);
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore old versions use servicePath, new versions serviceDir. Types will use only one of them
+    this.serviceDirPath = this.serverless.config.serviceDir || this.serverless.config.servicePath;
+    this.workDirPath = path.join(this.serviceDirPath, WORK_FOLDER);
     this.buildDirPath = path.join(this.workDirPath, BUILD_FOLDER);
-
-    const withDefaultOptions = mergeRight(DEFAULT_BUILD_OPTIONS);
-    this.buildOptions = withDefaultOptions<Configuration>(
-      this.serverless.service.custom?.esbuild ?? {}
-    );
 
     this.hooks = {
       'before:run:run': async () => {
@@ -86,14 +102,14 @@ export class EsbuildPlugin implements Plugin {
         await this.copyExtras();
       },
       'before:offline:start': async () => {
-        await this.bundle();
+        await this.bundle(true);
         await this.packExternalModules();
         await this.copyExtras();
         await this.preOffline();
         this.watch();
       },
       'before:offline:start:init': async () => {
-        await this.bundle();
+        await this.bundle(true);
         await this.packExternalModules();
         await this.copyExtras();
         await this.preOffline();
@@ -126,24 +142,70 @@ export class EsbuildPlugin implements Plugin {
     };
   }
 
+  /**
+   * Checks if the runtime for the given function is nodejs.
+   * If the runtime is not set , checks the global runtime.
+   * @param {Serverless.FunctionDefinitionHandler} func the function to be checked
+   * @returns {boolean} true if the function/global runtime is nodejs; false, otherwise
+   */
+  private isNodeFunction(func: Serverless.FunctionDefinitionHandler): boolean {
+    const runtime = func.runtime || this.serverless.service.provider.runtime;
+    const runtimeMatcher = providerRuntimeMatcher[this.serverless.service.provider.name];
+    return Boolean(runtimeMatcher?.[runtime]);
+  }
+
   get functions(): Record<string, Serverless.FunctionDefinitionHandler> {
     if (this.options.function) {
-      return {
-        [this.options.function]: this.serverless.service.getFunction(
-          this.options.function
-        ) as Serverless.FunctionDefinitionHandler,
-      };
+      //only return the function if it's a node function:
+      const func = this.serverless.service.getFunction(this.options.function) as Serverless.FunctionDefinitionHandler;
+      return this.isNodeFunction(func) ? { [this.options.function]: func } : {};
     }
 
-    return this.serverless.service.functions as Record<
-      string,
-      Serverless.FunctionDefinitionHandler
-    >;
+    // ignore all functions with a different runtime than nodejs:
+    const nodeFunctions: Record<string, Serverless.FunctionDefinitionHandler> = {};
+    const functions = this.serverless.service.functions;
+    for(const funcName in functions) {
+      const func = functions[funcName] as Serverless.FunctionDefinitionHandler;
+      if (this.isNodeFunction(func)) {
+        nodeFunctions[funcName] = func;
+      }
+    }
+    return nodeFunctions;
+  }
+
+  get plugins(): Plugins {
+    if (!this.buildOptions.plugins) return;
+
+    const plugins: Plugins | ReturnPluginsFn = require(path.join(
+      this.serviceDirPath,
+      this.buildOptions.plugins
+    ));
+
+    if (typeof plugins === 'function') {
+      return plugins(this.serverless);
+    }
+
+    return plugins;
+  }
+
+  private getCachedOptions = memoizeWith(always('cache'), () => {
+    const runtimeMatcher = providerRuntimeMatcher[this.serverless.service.provider.name];
+    const target = runtimeMatcher?.[this.serverless.service.provider.runtime];
+    const resolvedOptions = {
+      ...(target ? { target } : {}),
+    };
+    const withDefaultOptions = mergeRight(DEFAULT_BUILD_OPTIONS);
+    const withResolvedOptions = mergeRight(withDefaultOptions(resolvedOptions));
+    return withResolvedOptions<Configuration>(this.serverless.service.custom?.esbuild ?? {});
+  });
+
+  get buildOptions() {
+    return this.getCachedOptions();
   }
 
   get rootFileNames() {
     return extractFileNames(
-      this.serverless.config.servicePath,
+      this.serviceDirPath,
       this.serverless.service.provider.name,
       this.functions
     );
@@ -169,60 +231,100 @@ export class EsbuildPlugin implements Plugin {
     fs.mkdirpSync(this.buildDirPath);
     fs.mkdirpSync(path.join(this.workDirPath, SERVERLESS_FOLDER));
     // exclude serverless-esbuild
+    this.serverless.service.package = {
+      ...(this.serverless.service.package || {}),
+      patterns: [
+        ...new Set([
+          ...(this.serverless.service.package?.include || []),
+          ...(this.serverless.service.package?.exclude || []).map(concat('!')),
+          ...(this.serverless.service.package?.patterns || []),
+          '!node_modules/serverless-esbuild',
+        ]),
+      ],
+    };
+
     for (const fnName in this.functions) {
       const fn = this.serverless.service.getFunction(fnName);
-      fn.package = fn.package || {
-        exclude: [],
-        include: [],
+      fn.package = {
+        ...(fn.package || {}),
+        patterns: [
+          ...new Set([
+            ...(fn.package?.include || []),
+            ...(fn.package?.exclude || []).map(concat('!')),
+            ...(fn.package?.patterns || []),
+          ]),
+        ],
       };
-
-      // Add plugin to excluded packages or an empty array if exclude is undefined
-      fn.package.exclude = [
-        ...new Set([...(fn.package.exclude || []), 'node_modules/serverless-esbuild']),
-      ];
     }
   }
 
   async bundle(incremental = false): Promise<BuildResult[]> {
     this.prepare();
-    this.serverless.cli.log('Compiling with esbuild...');
+    this.serverless.cli.log(`Compiling to ${this.buildOptions.target} bundle with esbuild...`);
 
-    return Promise.all(
-      this.rootFileNames.map(async ({ entry, func }) => {
-        const config: Omit<BuildOptions, 'watch'> = {
-          ...this.buildOptions,
-          external: [...this.buildOptions.external, ...this.buildOptions.exclude],
-          entryPoints: [entry],
-          outdir: path.join(this.buildDirPath, path.dirname(entry)),
-          platform: 'node',
-          incremental,
-        };
+    const bundleMapper = async (bundleInfo) => {
+      const { entry, func, functionAlias } = bundleInfo;
+      const config: Omit<BuildOptions, 'watch'> = {
+        ...this.buildOptions,
+        external: [
+          ...this.buildOptions.external,
+          ...(this.buildOptions.exclude === '*' || this.buildOptions.exclude.includes('*')
+            ? []
+            : this.buildOptions.exclude),
+        ],
+        entryPoints: [entry],
+        outdir: path.join(this.buildDirPath, path.dirname(entry)),
+        platform: 'node',
+        incremental,
+        plugins: this.plugins,
+      };
 
-        // esbuild v0.7.0 introduced config options validation, so I have to delete plugin specific options from esbuild config.
-        delete config['exclude'];
-        delete config['packager'];
-        delete config['packagePath'];
-        delete config['watch'];
+      // esbuild v0.7.0 introduced config options validation, so I have to delete plugin specific options from esbuild config.
+      delete config['concurrency'];
+      delete config['exclude'];
+      delete config['nativeZip'];
+      delete config['packager'];
+      delete config['packagePath'];
+      delete config['watch'];
+      delete config['keepOutputDirectory'];
+      delete config['packagerOptions'];
 
-        const result = await build(config);
+      const bundlePath = entry.substr(0, entry.lastIndexOf('.')) + '.js';
 
-        const bundlePath = entry.substr(0, entry.lastIndexOf('.')) + '.js';
-        return { result, bundlePath, func };
-      })
-    ).then(results => {
-      this.serverless.cli.log('Compiling completed.');
-      this.buildResults = results;
-      return results.map(r => r.result);
+      if (this.buildResults) {
+        const { result } = this.buildResults.find(({ func: fn }) => fn.name === func.name);
+        await result.rebuild();
+        return { result, bundlePath, func, functionAlias };
+      }
+
+      const result = await build(config);
+
+      if (config.metafile) {
+        fs.writeFileSync(
+          path.join(this.buildDirPath, `${trimExtension(entry)}-meta.json`),
+          JSON.stringify(result.metafile, null, 2)
+        );
+      }
+
+      return { result, bundlePath, func, functionAlias };
+    };
+    this.serverless.cli.log(
+      `Compiling with concurrency: ${this.buildOptions.concurrency ?? 'Infinity'}`
+    );
+    this.buildResults = await pMap(this.rootFileNames, bundleMapper, {
+      concurrency: this.buildOptions.concurrency,
     });
+    this.serverless.cli.log('Compiling completed.');
+    return this.buildResults.map((r) => r.result);
   }
 
-  /** Link or copy extras such as node_modules or package.include definitions */
+  /** Link or copy extras such as node_modules or package.patterns definitions */
   async copyExtras() {
     const { service } = this.serverless;
 
-    // include any "extras" from the "include" section
-    if (service.package.include && service.package.include.length > 0) {
-      const files = await globby(service.package.include);
+    // include any "extras" from the "patterns" section
+    if (service.package.patterns.length > 0) {
+      const files = await globby(service.package.patterns);
 
       for (const filename of files) {
         const destFileName = path.resolve(path.join(this.buildDirPath, filename));
@@ -233,7 +335,30 @@ export class EsbuildPlugin implements Plugin {
         }
 
         if (!fs.existsSync(destFileName)) {
-          fs.copySync(path.resolve(filename), path.resolve(path.join(this.buildDirPath, filename)));
+          fs.copySync(path.resolve(filename), destFileName);
+        }
+      }
+    }
+
+    // include any "extras" from the individual function "patterns" section
+    for (const fnName in this.functions) {
+      const fn = this.serverless.service.getFunction(fnName);
+      if (fn.package.patterns.length === 0) {
+        continue;
+      }
+      const files = await globby(fn.package.patterns);
+      for (const filename of files) {
+        const destFileName = path.resolve(
+          path.join(this.buildDirPath, `__only_${fn.name}`, filename)
+        );
+        const dirname = path.dirname(destFileName);
+
+        if (!fs.existsSync(dirname)) {
+          fs.mkdirpSync(dirname);
+        }
+
+        if (!fs.existsSync(destFileName)) {
+          fs.copySync(path.resolve(filename), destFileName);
         }
       }
     }
@@ -248,13 +373,13 @@ export class EsbuildPlugin implements Plugin {
 
     await fs.copy(
       path.join(this.workDirPath, SERVERLESS_FOLDER),
-      path.join(this.serverless.config.servicePath, SERVERLESS_FOLDER)
+      path.join(this.serviceDirPath, SERVERLESS_FOLDER)
     );
 
     if (this.options.function) {
       const fn = service.getFunction(this.options.function);
       fn.package.artifact = path.join(
-        this.serverless.config.servicePath,
+        this.serviceDirPath,
         SERVERLESS_FOLDER,
         path.basename(fn.package.artifact)
       );
@@ -263,9 +388,9 @@ export class EsbuildPlugin implements Plugin {
 
     if (service.package.individually) {
       const functionNames = service.getAllFunctions();
-      functionNames.forEach(name => {
+      functionNames.forEach((name) => {
         service.getFunction(name).package.artifact = path.join(
-          this.serverless.config.servicePath,
+          this.serviceDirPath,
           SERVERLESS_FOLDER,
           path.basename(service.getFunction(name).package.artifact)
         );
@@ -274,7 +399,7 @@ export class EsbuildPlugin implements Plugin {
     }
 
     service.package.artifact = path.join(
-      this.serverless.config.servicePath,
+      this.serviceDirPath,
       SERVERLESS_FOLDER,
       path.basename(service.package.artifact)
     );
@@ -283,8 +408,10 @@ export class EsbuildPlugin implements Plugin {
   async cleanup(): Promise<void> {
     await this.moveArtifacts();
     // Remove temp build folder
-    fs.removeSync(path.join(this.workDirPath));
+    if (!this.buildOptions.keepOutputDirectory) {
+      fs.removeSync(path.join(this.workDirPath));
+    }
   }
 }
 
-module.exports = EsbuildPlugin;
+module.exports = EsbuildServerlessPlugin;
